requirements.txt
pandas==2.0.3
python_dateutil==2.8.2
selenium==4.7.2
# pip install --upgrade google-api-python-client google-auth-httplib2 google-auth-oauthlib